# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Notification Service",
    "description": "Multi-channel (email/SMS/push) notification service with priority queueing, retries, idempotency, rate limiting, and delivery tracking.",
    "version": "1.0.0"
  },
  "paths": {
    "/notifications": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Create Notification",
        "operationId": "create_notification_notifications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/analytics/stats": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Channel Stats",
        "description": "Bonus: basic sent/failed counts by channel and status.",
        "operationId": "channel_stats_notifications_analytics_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/notifications/{notification_id}": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Get Notification",
        "operationId": "get_notification_notifications__notification_id__get",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/notifications": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User Notifications",
        "operationId": "get_user_notifications_users__user_id__notifications_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/preferences": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Preferences",
        "operationId": "get_preferences_users__user_id__preferences_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Set Preferences",
        "operationId": "set_preferences_users__user_id__preferences_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Channel": {
        "type": "string",
        "enum": [
          "email",
          "sms",
          "push"
        ],
        "title": "Channel"
      },
      "ChannelDeliveryStatus": {
        "type": "string",
        "enum": [
          "pending",
          "queued",
          "sent",
          "delivered",
          "failed",
          "skipped"
        ],
        "title": "ChannelDeliveryStatus"
      },
      "ChannelPreference": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/Channel"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "channel",
          "enabled"
        ],
        "title": "ChannelPreference"
      },
      "ChannelStatusResponse": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/Channel"
          },
          "status": {
            "$ref": "#/components/schemas/ChannelDeliveryStatus"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          }
        },
        "type": "object",
        "required": [
          "channel",
          "status",
          "attempt_count"
        ],
        "title": "ChannelStatusResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NotificationCreateRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "User Id"
          },
          "channels": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Channel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels",
            "description": "Channels to send on. If omitted, uses all channels the user is opted into."
          },
          "priority": {
            "$ref": "#/components/schemas/Priority",
            "default": "normal"
          },
          "template_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Name",
            "description": "Name of a stored template to use"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject",
            "description": "Used if no template_name given"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body",
            "description": "Used if no template_name given"
          },
          "variables": {
            "type": "object",
            "title": "Variables",
            "description": "Values substituted into {{placeholders}}"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key",
            "description": "Prevents duplicate sends on retried requests"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "NotificationCreateRequest"
      },
      "NotificationListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "page_size",
          "items"
        ],
        "title": "NotificationListResponse"
      },
      "NotificationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "status": {
            "$ref": "#/components/schemas/NotificationStatus"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelStatusResponse"
            },
            "type": "array",
            "title": "Channels",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "priority",
          "status",
          "idempotency_key",
          "created_at",
          "updated_at"
        ],
        "title": "NotificationResponse"
      },
      "NotificationStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "sent",
          "delivered",
          "failed",
          "partially_failed"
        ],
        "title": "NotificationStatus"
      },
      "PreferenceResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "preferences": {
            "items": {
              "$ref": "#/components/schemas/ChannelPreference"
            },
            "type": "array",
            "title": "Preferences"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "preferences"
        ],
        "title": "PreferenceResponse"
      },
      "PreferenceUpdateRequest": {
        "properties": {
          "preferences": {
            "items": {
              "$ref": "#/components/schemas/ChannelPreference"
            },
            "type": "array",
            "title": "Preferences"
          }
        },
        "type": "object",
        "required": [
          "preferences"
        ],
        "title": "PreferenceUpdateRequest"
      },
      "Priority": {
        "type": "string",
        "enum": [
          "critical",
          "high",
          "normal",
          "low"
        ],
        "title": "Priority"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}